Add Tyrian v2 which is rewritten and based on BS 4
Bootstrap 4 is a major rewrite of the Bootstrap project. The
new Tyrian version v2 is now based on Bootstrap 4 instead
of Bootstrap 3 as before. In line with this, Tyrian has been
rewritten in scss instead of less (just as Bootstrap).

Apart from minor changes, Tyrian v2 looks like Tyrian v1 for now.
A fix for a sticky footer has been included. Apart from that
its now possible to use flexbox (as it is supported by BS 4)

For further information, please have a look at:
https://getbootstrap.com/docs/4.0/migration/

Signed-off-by: Max Magorsch <arzano@gentoo.org>